+++ create.sh
@@ -12,6 +12,6 @@ docker rmi -f petrjahoda/zapsi_service:latest
 docker build -t petrjahoda/zapsi_service:latest .
 docker push petrjahoda/zapsi_service:latest
 
-docker rmi -f petrjahoda/zapsi_service:2020.1.3
-docker build -t petrjahoda/zapsi_service:2020.1.3 .
-docker push petrjahoda/zapsi_service:2020.1.3
+docker rmi -f petrjahoda/zapsi_service:2020.2.2
+docker build -t petrjahoda/zapsi_service:2020.2.2 .
+docker push petrjahoda/zapsi_service:2020.2.2